Move updateCodeLenses logic to AgentFixupControls (#5477)
Fixes https://github.com/sourcegraph/jetbrains/issues/2197.

## Test plan
Try to reproduce https://github.com/sourcegraph/jetbrains/issues/2197.

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

## Changelog
Move updateCodeLenses from agent.ts to AgentFixupControls.ts

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->
