+++ agent/src/AgentFixupControls.ts
@@ -1,10 +1,11 @@
+import type * as vscode from 'vscode'
 import type { QuickPickInput } from '../../vscode/src/edit/input/get-input'
 import type { FixupFile } from '../../vscode/src/non-stop/FixupFile'
 import type { FixupTask, FixupTaskID } from '../../vscode/src/non-stop/FixupTask'
 import { FixupCodeLenses } from '../../vscode/src/non-stop/codelenses/provider'
 import type { FixupActor, FixupFileCollection } from '../../vscode/src/non-stop/roles'
 import { type Agent, errorToCodyError } from './agent'
-import type { EditTask } from './protocol-alias'
+import type { EditTask, ProtocolCommand } from './protocol-alias'
 
 export class AgentFixupControls extends FixupCodeLenses {
     constructor(
@@ -59,6 +60,52 @@ export class AgentFixupControls extends FixupCodeLenses {
     visibleFilesWithTasksMaybeChanged(files: readonly FixupFile[]): void {}
 
     dispose() {}
+
+    override notifyCodeLensesChanged(uri: vscode.Uri, codeLenses: vscode.CodeLens[]) {
+        super.notifyCodeLensesChanged(uri, codeLenses)
+        void this.updateCodeLenses(uri, codeLenses)
+    }
+
+    private async updateCodeLenses(uri: vscode.Uri, codeLenses: vscode.CodeLens[]): Promise<void> {
+        // VS Code supports icons in code lenses, but we cannot render these through agent.
+        // We need to strip any icons from the title and provide those seperately, so the client can decide how to render them.
+        const agentLenses = codeLenses.map(lens => ({
+            ...lens,
+            command: lens.command
+                ? {
+                      ...lens.command,
+                      title: this.splitIconsFromTitle(lens.command.title),
+                  }
+                : undefined,
+        }))
+
+        this.notify('codeLenses/display', {
+            uri: uri.toString(),
+            codeLenses: agentLenses,
+        })
+    }
+
+    /**
+     * Matches VS Code codicon syntax, e.g. $(cody-logo)
+     * Source: https://sourcegraph.com/github.com/microsoft/vscode@f34d4/-/blob/src/vs/base/browser/ui/iconLabel/iconLabels.ts?L9
+     */
+    private labelWithIconsRegex = /(\\)?\$\(([A-Za-z0-9-]+(?:~[A-Za-z]+)?)\)/g
+    /**
+     * Given a title, such as "$(cody-logo) Cody", returns the raw
+     * title without icons and the icons matched with their respective positions.
+     */
+    private splitIconsFromTitle(title: string): ProtocolCommand['title'] {
+        const icons: { value: string; position: number }[] = []
+        const matches = [...title.matchAll(this.labelWithIconsRegex)]
+
+        for (const match of matches) {
+            if (match.index !== undefined) {
+                icons.push({ value: match[0], position: match.index })
+            }
+        }
+
+        return { text: title.replace(this.labelWithIconsRegex, ''), icons }
+    }
 
     public static serialize(task: FixupTask): EditTask {
         return {
